Introduce listRelevantDevPortalDevices.
