+++ devices_test.go
@@ -1,0 +1,212 @@
+package main
+
+import (
+	"fmt"
+	"net/http"
+	"testing"
+
+	"github.com/bitrise-io/go-xcode/devportalservice"
+	"github.com/bitrise-steplib/steps-ios-auto-provision-appstoreconnect/appstoreconnect"
+	"github.com/bitrise-steplib/steps-ios-auto-provision-appstoreconnect/autoprovision"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
+)
+
+func Test_registerMissingDevices_alreadyRegistered(t *testing.T) {
+	mockClient := &MockClientRegisterDevice{}
+	successClient := appstoreconnect.NewClient(mockClient, "keyID", "issueID", []byte("privateKey"))
+
+	args := struct {
+		client           *appstoreconnect.Client
+		bitriseDevices   []devportalservice.TestDevice
+		devportalDevices []appstoreconnect.Device
+	}{
+		client: successClient,
+		bitriseDevices: []devportalservice.TestDevice{{
+			DeviceID:   "71153a920968f2842d360",
+			DeviceType: "ios",
+		}},
+		devportalDevices: []appstoreconnect.Device{{
+			Attributes: appstoreconnect.DeviceAttributes{
+				UDID: "71153a920968f2842d360",
+			},
+			ID: "12",
+		}},
+	}
+	want := []appstoreconnect.Device(nil)
+
+	got, err := registerMissingTestDevices(args.client, args.bitriseDevices, args.devportalDevices)
+
+	require.NoError(t, err, "registerMissingDevices() error")
+	require.Equal(t, want, got, "registerMissingDevices()")
+	mockClient.AssertNotCalled(t, "PostDevice")
+	mockClient.AssertExpectations(t)
+}
+
+func Test_registerMissingDevices_newDevice(t *testing.T) {
+	mockClient := &MockClientRegisterDevice{}
+	mockClient.
+		On("PostDevice", mock.AnythingOfType("*http.Request")).
+		Return(newResponse(t, http.StatusOK,
+			map[string]interface{}{
+				"data": map[string]interface{}{
+					"id": "12",
+					"attributes": map[string]interface{}{
+						"deviceClass": "IPHONE",
+					},
+				},
+			},
+		), nil)
+	successClient := appstoreconnect.NewClient(mockClient, "keyID", "issueID", []byte("privateKey"))
+
+	args := struct {
+		client           *appstoreconnect.Client
+		bitriseDevices   []devportalservice.TestDevice
+		devportalDevices []appstoreconnect.Device
+	}{
+		client: successClient,
+		bitriseDevices: []devportalservice.TestDevice{{
+			DeviceID:   "71153a920968f2842d360",
+			DeviceType: "ios",
+		}},
+		devportalDevices: []appstoreconnect.Device{},
+	}
+	want := []appstoreconnect.Device{{
+		Attributes: appstoreconnect.DeviceAttributes{
+			DeviceClass: appstoreconnect.Iphone,
+		},
+		ID: "12",
+	}}
+
+	got, err := registerMissingTestDevices(args.client, args.bitriseDevices, args.devportalDevices)
+
+	require.NoError(t, err, "registerMissingDevices()")
+	require.Equal(t, want, got, "registerMissingDevices()")
+	mockClient.AssertExpectations(t)
+}
+
+func Test_registerMissingDevices_invalidUDID(t *testing.T) {
+	mockClient := &MockClientRegisterDevice{}
+	mockClient.
+		On("PostDevice", mock.AnythingOfType("*http.Request")).
+		Return(newResponse(t, http.StatusConflict,
+			map[string]interface{}{
+				"errors": []interface{}{map[string]interface{}{"detail": "ENTITY_ERROR.ATTRIBUTE.INVALID: An attribute in the provided entity has invalid value: An invalid value 'xxx' was provided for the parameter 'udid'."}},
+			}), nil)
+	failureClient := appstoreconnect.NewClient(mockClient, "keyID", "issueID", []byte("privateKey"))
+
+	args := struct {
+		client           *appstoreconnect.Client
+		bitriseDevices   []devportalservice.TestDevice
+		devportalDevices []appstoreconnect.Device
+	}{
+		client: failureClient,
+		bitriseDevices: []devportalservice.TestDevice{
+			{
+				DeviceID:   "invalid-udid",
+				DeviceType: "ios",
+			},
+			{
+				DeviceID:   "71153a920968f2842d360",
+				DeviceType: "ios",
+			},
+		},
+		devportalDevices: []appstoreconnect.Device{{
+			Attributes: appstoreconnect.DeviceAttributes{
+				UDID: "71153a920968f2842d360",
+			},
+			ID: "12",
+		}},
+	}
+	want := []appstoreconnect.Device(nil)
+
+	got, err := registerMissingTestDevices(args.client, args.bitriseDevices, args.devportalDevices)
+	require.NoError(t, err, "registerMissingDevices()")
+	require.Equal(t, want, got, "registerMissingDevices()")
+	mockClient.AssertExpectations(t)
+}
+
+func Test_listRelevantDevPortalDevices_filtersDevicesForPlatform(t *testing.T) {
+	tests := []struct {
+		platform      autoprovision.Platform
+		deviceClass   appstoreconnect.DeviceClass
+		devicesLength int
+	}{
+		{platform: autoprovision.IOS, deviceClass: appstoreconnect.AppleWatch, devicesLength: 1},
+		{platform: autoprovision.IOS, deviceClass: appstoreconnect.Ipad, devicesLength: 1},
+		{platform: autoprovision.IOS, deviceClass: appstoreconnect.Iphone, devicesLength: 1},
+		{platform: autoprovision.IOS, deviceClass: appstoreconnect.Ipod, devicesLength: 1},
+		{platform: autoprovision.IOS, deviceClass: appstoreconnect.AppleTV, devicesLength: 0},
+		{platform: autoprovision.IOS, deviceClass: appstoreconnect.Mac, devicesLength: 0},
+
+		{platform: autoprovision.TVOS, deviceClass: appstoreconnect.AppleWatch, devicesLength: 0},
+		{platform: autoprovision.TVOS, deviceClass: appstoreconnect.Ipad, devicesLength: 0},
+		{platform: autoprovision.TVOS, deviceClass: appstoreconnect.Iphone, devicesLength: 0},
+		{platform: autoprovision.TVOS, deviceClass: appstoreconnect.Ipod, devicesLength: 0},
+		{platform: autoprovision.TVOS, deviceClass: appstoreconnect.AppleTV, devicesLength: 1},
+		{platform: autoprovision.TVOS, deviceClass: appstoreconnect.Mac, devicesLength: 0},
+
+		{platform: autoprovision.MacOS, deviceClass: appstoreconnect.AppleWatch, devicesLength: 0},
+		{platform: autoprovision.MacOS, deviceClass: appstoreconnect.Ipad, devicesLength: 0},
+		{platform: autoprovision.MacOS, deviceClass: appstoreconnect.Iphone, devicesLength: 0},
+		{platform: autoprovision.MacOS, deviceClass: appstoreconnect.Ipod, devicesLength: 0},
+		{platform: autoprovision.MacOS, deviceClass: appstoreconnect.AppleTV, devicesLength: 0},
+		{platform: autoprovision.MacOS, deviceClass: appstoreconnect.Mac, devicesLength: 0},
+	}
+
+	for _, tt := range tests {
+		t.Run(fmt.Sprintf("Given platform is %s and device is %s when missing devices are registered then valid devices length should be %d", tt.platform, tt.deviceClass, tt.devicesLength), func(t *testing.T) {
+			client := givenGetDevicesReturnsDeviceWithClass(t, tt.deviceClass)
+
+			got, err := listRelevantDevPortalDevices(client, nil, tt.platform)
+			assert.NoError(t, err)
+			assert.Equal(t, tt.devicesLength, len(got))
+		})
+	}
+}
+
+type MockClientRegisterDevice struct {
+	mock.Mock
+}
+
+func (c *MockClientRegisterDevice) Do(req *http.Request) (*http.Response, error) {
+	fmt.Printf("do called: %#v - %#v\n", req.Method, req.URL.Path)
+
+	switch {
+	case req.URL.Path == "/v1/devices" && req.Method == "POST":
+		return c.PostDevice(req)
+	case req.URL.Path == "/v1/devices" && req.Method == "GET":
+		return c.GetDevices(req)
+	}
+
+	return nil, fmt.Errorf("invalid endpoint called: %s, method: %s", req.URL.Path, req.Method)
+}
+
+func (c *MockClientRegisterDevice) PostDevice(req *http.Request) (*http.Response, error) {
+	args := c.Called(req)
+	return args.Get(0).(*http.Response), args.Error(1)
+}
+
+func (c *MockClientRegisterDevice) GetDevices(req *http.Request) (*http.Response, error) {
+	args := c.Called(req)
+	return args.Get(0).(*http.Response), args.Error(1)
+}
+
+func givenGetDevicesReturnsDeviceWithClass(t *testing.T, testDeviceClass appstoreconnect.DeviceClass) *appstoreconnect.Client {
+	mockClient := &MockClientRegisterDevice{}
+	mockClient.
+		On("GetDevices", mock.AnythingOfType("*http.Request")).
+		Return(newResponse(t, http.StatusOK, map[string]interface{}{
+			"data": []map[string]interface{}{
+				{
+					"id": "1",
+					"attributes": map[string]interface{}{
+						"deviceClass": testDeviceClass,
+					},
+				},
+			},
+		},
+		), nil)
+	return appstoreconnect.NewClient(mockClient, "keyID", "issueID", []byte("privateKey"))
+}
